split format and push
